started working on charging IC
can't for the life of me tell if it will work with low battery
voltage... seems like it might work, but it won't start switching until
the battery voltage is above 3.3V, in which case it acts as a buck
regulator to 3.3V. That 3.3V is adjustable, however, so if we can get
away with 3.0V on the other IC's which we need to power, we might be
okay... It will boost the battery voltage up to 5V if a usb device is
plugged in, but I don't think that will be useful for any of the circuit
elements.  if you get a chance, can you read over the data sheet also?
It is a complicated device.