I2C - BROKEN...
gets stuck in an infinite loop where the write function gets called over
and over. If I single step then everything works as intended. If I let
the chip run, then the I2C LED's never turn on. The older code worked,
but I don't like the way it was written. It really needs to be
non-blocking.